--- a/src/Razor/src/Microsoft.AspNetCore.Razor.ProjectEngineHost/CompilationTagHelperResolver.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.ProjectEngineHost/CompilationTagHelperResolver.cs
@@ -13,6 +13,7 @@ using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.PooledObjects;
 using Microsoft.AspNetCore.Razor.Serialization;
 using Microsoft.AspNetCore.Razor.Telemetry;
+using Microsoft.AspNetCore.Razor.Utilities;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Razor;
 
@@ -43,7 +44,7 @@ internal class CompilationTagHelperResolver(ITelemetryReporter? telemetryReporte
             return TagHelperResolutionResult.Empty;
         }
 
-        var results = new HashSet<TagHelperDescriptor>();
+        var results = new HashSet<TagHelperDescriptor>(TagHelperChecksumComparer.Instance);
         var context = TagHelperDescriptorProviderContext.Create(results);
         context.ExcludeHidden = true;
         context.IncludeDocumentation = true;