--- a/src/Razor/src/Microsoft.AspNetCore.Razor.ProjectEngineHost/Serialization/TagHelperDeltaResult.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.ProjectEngineHost/Serialization/TagHelperDeltaResult.cs
@@ -4,6 +4,12 @@
 using System.Collections.Immutable;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.PooledObjects;
+using Microsoft.AspNetCore.Razor.Utilities;
+
+#if DEBUG
+using System.Collections.Generic;
+using System.Diagnostics;
+#endif
 
 namespace Microsoft.AspNetCore.Razor.Serialization;
 
@@ -25,18 +31,22 @@ internal record TagHelperDeltaResult(
         //
         // 1. This TagHelperDeltaResult.Apply where we don't iterate / Contains check the "base" collection.
         // 2. The rest of the Razor project system. Everything there is always indexed / iterated as a list.
-        using var _ = ArrayBuilderPool<TagHelperDescriptor>.GetPooledObject(out var newTagHelpers);
-        newTagHelpers.SetCapacityIfLarger(baseTagHelpers.Length + Added.Length - Removed.Length);
-        newTagHelpers.AddRange(Added);
+        using var _ = ArrayBuilderPool<TagHelperDescriptor>.GetPooledObject(out var result);
+        result.SetCapacityIfLarger(baseTagHelpers.Length + Added.Length - Removed.Length);
 
-        foreach (var existingTagHelper in baseTagHelpers)
+        result.AddRange(Added);
+        result.AddRange(TagHelperDelta.Compute(Removed, baseTagHelpers));
+
+#if DEBUG
+        // Ensure that there are no duplicate tag helpers in the result.
+        var set = new HashSet<TagHelperDescriptor>(TagHelperChecksumComparer.Instance);
+
+        foreach (var item in result)
         {
-            if (!Removed.Contains(existingTagHelper))
-            {
-                newTagHelpers.Add(existingTagHelper);
-            }
+            Debug.Assert(set.Add(item), $"{nameof(TagHelperDeltaResult)}.{nameof(Apply)} should not contain any duplicates!");
         }
+#endif
 
-        return newTagHelpers.ToImmutable();
+        return result.DrainToImmutable();
     }
 }