--- a/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/IRemoteTagHelperProviderService.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/IRemoteTagHelperProviderService.cs
@@ -19,7 +19,7 @@ internal interface IRemoteTagHelperProviderService
     ValueTask<TagHelperDeltaResult> GetTagHelpersDeltaAsync(
         RazorPinnedSolutionInfoWrapper solutionInfo,
         ProjectSnapshotHandle projectHandle,
-        string? factoryTypeName,
+        string factoryTypeName,
         int lastResultId,
         CancellationToken cancellationToken);
 }