--- a/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/OOPTagHelperResolver.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/OOPTagHelperResolver.cs
@@ -59,6 +59,7 @@ internal class OOPTagHelperResolver : ITagHelperResolver
         try
         {
             TagHelperResolutionResult? result = null;
+
             if (factory != null)
             {
                 result = await ResolveTagHelpersOutOfProcessAsync(factory, workspaceProject, projectSnapshot, cancellationToken).ConfigureAwait(false);
@@ -81,7 +82,11 @@ internal class OOPTagHelperResolver : ITagHelperResolver
         // when it's disconnected (user stops the process).
         //
         // This will change in the future to an easier to consume API but for VS RTM this is what we have.
-        var remoteClient = await RazorRemoteHostClient.TryGetClientAsync(_workspace.Services, RazorServiceDescriptors.TagHelperProviderServiceDescriptors, RazorRemoteServiceCallbackDispatcherRegistry.Empty, cancellationToken);
+        var remoteClient = await RazorRemoteHostClient.TryGetClientAsync(
+            _workspace.Services,
+            RazorServiceDescriptors.TagHelperProviderServiceDescriptors,
+            RazorRemoteServiceCallbackDispatcherRegistry.Empty,
+            cancellationToken);
 
         if (remoteClient is null)
         {
@@ -95,11 +100,13 @@ internal class OOPTagHelperResolver : ITagHelperResolver
         }
 
         var projectHandle = new ProjectSnapshotHandle(workspaceProject.Id, projectSnapshot.Configuration, projectSnapshot.RootNamespace);
+        var factoryTypeName = factory.GetType().AssemblyQualifiedName;
+
         var result = await remoteClient.TryInvokeAsync<IRemoteTagHelperProviderService, TagHelperDeltaResult>(
             workspaceProject.Solution,
-            (service, solutionInfo, innerCancellationToken) => service.GetTagHelpersDeltaAsync(solutionInfo, projectHandle, factory?.GetType().AssemblyQualifiedName, lastResultId, innerCancellationToken),
-            cancellationToken
-        );
+            (service, solutionInfo, innerCancellationToken) =>
+                service.GetTagHelpersDeltaAsync(solutionInfo, projectHandle, factoryTypeName, lastResultId, innerCancellationToken),
+            cancellationToken);
 
         if (!result.HasValue)
         {