--- a/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/RazorServiceBase.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/RazorServiceBase.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.AspNetCore.Razor.Telemetry;
-using Microsoft.CodeAnalysis.Razor;
 using Microsoft.ServiceHub.Framework;
 
 namespace Microsoft.CodeAnalysis.Remote.Razor;
@@ -12,16 +10,12 @@ internal abstract class RazorServiceBase : IDisposable
 {
     protected readonly ServiceBrokerClient ServiceBrokerClient;
 
-    public RazorServiceBase(IServiceBroker serviceBroker, ITelemetryReporter telemetryReporter)
+    public RazorServiceBase(IServiceBroker serviceBroker)
     {
-        RazorServices = new RazorServices(telemetryReporter);
-
 #pragma warning disable VSTHRD012 // Provide JoinableTaskFactory where allowed
         ServiceBrokerClient = new ServiceBrokerClient(serviceBroker);
 #pragma warning restore
     }
-
-    protected RazorServices RazorServices { get; }
 
     public void Dispose()
     {