--- a/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/RemoteTagHelperDeltaProvider.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/RemoteTagHelperDeltaProvider.cs
@@ -3,21 +3,16 @@
 
 using System.Collections.Immutable;
 using Microsoft.AspNetCore.Razor.Language;
-using Microsoft.AspNetCore.Razor.PooledObjects;
 using Microsoft.AspNetCore.Razor.Serialization;
+using Microsoft.AspNetCore.Razor.Utilities;
 
 namespace Microsoft.CodeAnalysis.Remote.Razor;
 
 internal class RemoteTagHelperDeltaProvider
 {
-    private readonly TagHelperResultCache _resultCache;
-    private readonly object _resultIdLock = new();
+    private readonly TagHelperResultCache _resultCache = new();
+    private readonly object _gate = new();
     private int _currentResultId;
-
-    public RemoteTagHelperDeltaProvider()
-    {
-        _resultCache = new TagHelperResultCache();
-    }
 
     public TagHelperDeltaResult GetTagHelpersDelta(
         ProjectId projectId,
@@ -25,15 +20,16 @@ internal class RemoteTagHelperDeltaProvider
         ImmutableArray<TagHelperDescriptor> currentTagHelpers)
     {
         var cacheHit = _resultCache.TryGet(projectId, lastResultId, out var cachedTagHelpers);
+
         if (!cacheHit)
         {
             cachedTagHelpers = ImmutableArray<TagHelperDescriptor>.Empty;
         }
 
-        var added = GetAddedTagHelpers(currentTagHelpers, cachedTagHelpers!);
-        var removed = GetRemovedTagHelpers(currentTagHelpers, cachedTagHelpers!);
+        var added = TagHelperDelta.Compute(cachedTagHelpers, currentTagHelpers);
+        var removed = TagHelperDelta.Compute(currentTagHelpers, cachedTagHelpers);
 
-        lock (_resultIdLock)
+        lock (_gate)
         {
             var resultId = _currentResultId;
             if (added.Length > 0 || removed.Length > 0)
@@ -51,59 +47,5 @@ internal class RemoteTagHelperDeltaProvider
             var result = new TagHelperDeltaResult(cacheHit, resultId, added, removed);
             return result;
         }
-    }
-
-    private static ImmutableArray<TagHelperDescriptor> GetAddedTagHelpers(ImmutableArray<TagHelperDescriptor> current, ImmutableArray<TagHelperDescriptor> old)
-    {
-        if (old.Length == 0)
-        {
-            // Everything is considered added when there is no collection to compare to.
-            return current;
-        }
-
-        if (current.Length == 0)
-        {
-            // No new descriptors so can't possibly add any
-            return ImmutableArray<TagHelperDescriptor>.Empty;
-        }
-
-        using var _ = ArrayBuilderPool<TagHelperDescriptor>.GetPooledObject(out var added);
-
-        foreach (var tagHelper in current)
-        {
-            if (!old.Contains(tagHelper))
-            {
-                added.Add(tagHelper);
-            }
-        }
-
-        return added.ToImmutable();
-    }
-
-    private static ImmutableArray<TagHelperDescriptor> GetRemovedTagHelpers(ImmutableArray<TagHelperDescriptor> current, ImmutableArray<TagHelperDescriptor> old)
-    {
-        if (old.Length == 0)
-        {
-            // Can't have anything removed if there's nothing to compare to
-            return ImmutableArray<TagHelperDescriptor>.Empty;
-        }
-
-        if (current.Length == 0)
-        {
-            // Current collection has nothing so anything in "old" must have been removed
-            return old;
-        }
-
-        using var _ = ArrayBuilderPool<TagHelperDescriptor>.GetPooledObject(out var removed);
-
-        foreach (var tagHelper in old)
-        {
-            if (!current.Contains(tagHelper))
-            {
-                removed.Add(tagHelper);
-            }
-        }
-
-        return removed.ToImmutable();
     }
 }