--- a/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/RemoteTagHelperProviderService.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/RemoteTagHelperProviderService.cs
@@ -25,13 +25,26 @@ internal sealed class RemoteTagHelperProviderService : RazorServiceBase, IRemote
         _tagHelperDeltaProvider = new RemoteTagHelperDeltaProvider();
     }
 
-    public ValueTask<TagHelperResolutionResult> GetTagHelpersAsync(RazorPinnedSolutionInfoWrapper solutionInfo, ProjectSnapshotHandle projectHandle, string factoryTypeName, CancellationToken cancellationToken = default)
-        => RazorBrokeredServiceImplementation.RunServiceAsync(cancellationToken => GetTagHelpersCoreAsync(solutionInfo, projectHandle, factoryTypeName, cancellationToken), cancellationToken);
+    public ValueTask<TagHelperResolutionResult> GetTagHelpersAsync(
+        RazorPinnedSolutionInfoWrapper solutionInfo,
+        ProjectSnapshotHandle projectHandle,
+        string factoryTypeName,
+        CancellationToken cancellationToken)
+        => RazorBrokeredServiceImplementation.RunServiceAsync(
+            cancellationToken => GetTagHelpersCoreAsync(solutionInfo, projectHandle, factoryTypeName, cancellationToken),
+            cancellationToken);
 
-    public ValueTask<TagHelperDeltaResult> GetTagHelpersDeltaAsync(RazorPinnedSolutionInfoWrapper solutionInfo, ProjectSnapshotHandle projectHandle, string? factoryTypeName, int lastResultId, CancellationToken cancellationToken)
-        => RazorBrokeredServiceImplementation.RunServiceAsync(cancellationToken => GetTagHelpersDeltaCoreAsync(solutionInfo, projectHandle, factoryTypeName, lastResultId, cancellationToken), cancellationToken);
+    public ValueTask<TagHelperDeltaResult> GetTagHelpersDeltaAsync(
+        RazorPinnedSolutionInfoWrapper solutionInfo,
+        ProjectSnapshotHandle projectHandle,
+        string factoryTypeName,
+        int lastResultId,
+        CancellationToken cancellationToken)
+        => RazorBrokeredServiceImplementation.RunServiceAsync(
+            cancellationToken => GetTagHelpersDeltaCoreAsync(solutionInfo, projectHandle, factoryTypeName, lastResultId, cancellationToken),
+            cancellationToken);
 
-    private async ValueTask<TagHelperResolutionResult> GetTagHelpersCoreAsync(RazorPinnedSolutionInfoWrapper solutionInfo, ProjectSnapshotHandle projectHandle, string? factoryTypeName, CancellationToken cancellationToken)
+    private async ValueTask<TagHelperResolutionResult> GetTagHelpersCoreAsync(RazorPinnedSolutionInfoWrapper solutionInfo, ProjectSnapshotHandle projectHandle, string factoryTypeName, CancellationToken cancellationToken)
     {
         if (projectHandle is null)
         {
@@ -57,7 +70,7 @@ internal sealed class RemoteTagHelperProviderService : RazorServiceBase, IRemote
         return await _tagHelperResolver.GetTagHelpersAsync(workspaceProject, projectHandle.Configuration, factoryTypeName, cancellationToken).ConfigureAwait(false);
     }
 
-    public async ValueTask<TagHelperDeltaResult> GetTagHelpersDeltaCoreAsync(RazorPinnedSolutionInfoWrapper solutionInfo, ProjectSnapshotHandle projectHandle, string? factoryTypeName, int lastResultId, CancellationToken cancellationToken)
+    public async ValueTask<TagHelperDeltaResult> GetTagHelpersDeltaCoreAsync(RazorPinnedSolutionInfoWrapper solutionInfo, ProjectSnapshotHandle projectHandle, string factoryTypeName, int lastResultId, CancellationToken cancellationToken)
     {
         var tagHelperResolutionResult = await GetTagHelpersCoreAsync(solutionInfo, projectHandle, factoryTypeName, cancellationToken).ConfigureAwait(false);
         var currentTagHelpers = tagHelperResolutionResult.Descriptors;