--- a/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/RemoteTagHelperProviderService.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/RemoteTagHelperProviderService.cs
@@ -31,7 +31,9 @@ internal sealed class RemoteTagHelperProviderService : RazorServiceBase, IRemote
         string factoryTypeName,
         CancellationToken cancellationToken)
         => RazorBrokeredServiceImplementation.RunServiceAsync(
-            cancellationToken => GetTagHelpersCoreAsync(solutionInfo, projectHandle, factoryTypeName, cancellationToken),
+            solutionInfo,
+            ServiceBrokerClient,
+            solution => GetTagHelpersCoreAsync(solution, projectHandle, factoryTypeName, cancellationToken),
             cancellationToken);
 
     public ValueTask<TagHelperDeltaResult> GetTagHelpersDeltaAsync(
@@ -41,38 +43,28 @@ internal sealed class RemoteTagHelperProviderService : RazorServiceBase, IRemote
         int lastResultId,
         CancellationToken cancellationToken)
         => RazorBrokeredServiceImplementation.RunServiceAsync(
-            cancellationToken => GetTagHelpersDeltaCoreAsync(solutionInfo, projectHandle, factoryTypeName, lastResultId, cancellationToken),
+            solutionInfo,
+            ServiceBrokerClient,
+            solution => GetTagHelpersDeltaCoreAsync(solution, projectHandle, factoryTypeName, lastResultId, cancellationToken),
             cancellationToken);
 
-    private async ValueTask<TagHelperResolutionResult> GetTagHelpersCoreAsync(RazorPinnedSolutionInfoWrapper solutionInfo, ProjectSnapshotHandle projectHandle, string factoryTypeName, CancellationToken cancellationToken)
+    private ValueTask<TagHelperResolutionResult> GetTagHelpersCoreAsync(
+        Solution solution,
+        ProjectSnapshotHandle projectHandle,
+        string factoryTypeName,
+        CancellationToken cancellationToken)
+        => solution.GetProject(projectHandle.ProjectId) is Project project
+            ? _tagHelperResolver.GetTagHelpersAsync(project, projectHandle.Configuration, factoryTypeName, cancellationToken)
+            : new(TagHelperResolutionResult.Empty);
+
+    private async ValueTask<TagHelperDeltaResult> GetTagHelpersDeltaCoreAsync(
+        Solution solution,
+        ProjectSnapshotHandle projectHandle,
+        string factoryTypeName,
+        int lastResultId,
+        CancellationToken cancellationToken)
     {
-        if (projectHandle is null)
-        {
-            throw new ArgumentNullException(nameof(projectHandle));
-        }
-
-        if (string.IsNullOrEmpty(factoryTypeName))
-        {
-            throw new ArgumentException($"'{nameof(factoryTypeName)}' cannot be null or empty.", nameof(factoryTypeName));
-        }
-
-        // We should replace the below call: https://github.com/dotnet/razor-tooling/issues/6316
-#pragma warning disable CS0618 // Type or member is obsolete
-        var solution = await solutionInfo.GetSolutionAsync(ServiceBrokerClient, cancellationToken).ConfigureAwait(false);
-#pragma warning restore CS0618 // Type or member is obsolete
-        var workspaceProject = solution.GetProject(projectHandle.ProjectId);
-
-        if (workspaceProject is null)
-        {
-            return TagHelperResolutionResult.Empty;
-        }
-
-        return await _tagHelperResolver.GetTagHelpersAsync(workspaceProject, projectHandle.Configuration, factoryTypeName, cancellationToken).ConfigureAwait(false);
-    }
-
-    public async ValueTask<TagHelperDeltaResult> GetTagHelpersDeltaCoreAsync(RazorPinnedSolutionInfoWrapper solutionInfo, ProjectSnapshotHandle projectHandle, string factoryTypeName, int lastResultId, CancellationToken cancellationToken)
-    {
-        var tagHelperResolutionResult = await GetTagHelpersCoreAsync(solutionInfo, projectHandle, factoryTypeName, cancellationToken).ConfigureAwait(false);
+        var tagHelperResolutionResult = await GetTagHelpersCoreAsync(solution, projectHandle, factoryTypeName, cancellationToken).ConfigureAwait(false);
         var currentTagHelpers = tagHelperResolutionResult.Descriptors;
 
         return _tagHelperDeltaProvider.GetTagHelpersDelta(projectHandle.ProjectId, lastResultId, currentTagHelpers);