Fixes #6316
Fixes #8273

This pull request covers a few things:

1. A general clean up and refactoring of the code that runs in Roslyn OOP in preparation for other changes I'll be making there to avoid serializing tag helpers when unnecessary.
2. A fix for #6316 to use a better non-deprecated API provided by Roslyn.
3. A huge improvement in the performance of tag helper delta calculations. (Roughly 30-50% fewer allocations and as much as a 20x performance improvement!)

The latter piece is something I've been intending to do for a long while, and tag helper checksums make this much easier.

Here's the `RemoteTagHelperDeltaProviderBenchmark` results.

### Old (main)

|                                        Method |        Job |            Toolchain |         Mean |      Error |       StdDev |       Median |           Min |          Max | Allocated |
|---------------------------------------------- |----------- |--------------------- |-------------:|-----------:|-------------:|-------------:|--------------:|-------------:|----------:|
|               &#39;Calculate Delta - New project&#39; | Job-CUDZQP |             .NET 7.0 |     20.14 μs |   0.863 μs |     2.448 μs |     20.00 μs |     16.100 μs |     27.10 μs |         - |
|            &#39;Calculate Delta - Remove project&#39; | Job-CUDZQP |             .NET 7.0 |     17.55 μs |   0.765 μs |     2.119 μs |     17.20 μs |     13.300 μs |     22.90 μs |         - |
|    &#39;Calculate Delta - Add lots of TagHelpers&#39; | Job-CUDZQP |             .NET 7.0 | 33,130.71 μs | 659.613 μs | 1,737.686 μs | 33,119.45 μs | 28,079.450 μs | 37,551.25 μs |  326536 B |
| &#39;Calculate Delta - Remove lots of TagHelpers&#39; | Job-CUDZQP |             .NET 7.0 |  8,457.13 μs | 383.761 μs | 1,050.540 μs |  8,367.30 μs |  6,714.000 μs | 11,075.40 μs |  167752 B |
|     &#39;Calculate Delta - Mutate two TagHelpers&#39; | Job-CUDZQP |             .NET 7.0 | 18,794.13 μs | 647.050 μs | 1,897.685 μs | 18,783.40 μs | 14,468.400 μs | 23,527.20 μs |  307792 B |
|                 &#39;Calculate Delta - No change&#39; | Job-CUDZQP |             .NET 7.0 | 18,711.70 μs | 622.339 μs | 1,785.608 μs | 18,576.85 μs | 15,469.050 μs | 23,066.35 μs |  307984 B |
|               &#39;Calculate Delta - New project&#39; | Job-KWCLPV | .NET Framework 4.7.2 |     13.83 μs |   0.791 μs |     2.218 μs |     13.30 μs |      9.700 μs |     20.40 μs |         - |
|            &#39;Calculate Delta - Remove project&#39; | Job-KWCLPV | .NET Framework 4.7.2 |     11.16 μs |   0.714 μs |     2.014 μs |     10.45 μs |      7.900 μs |     17.90 μs |         - |
|    &#39;Calculate Delta - Add lots of TagHelpers&#39; | Job-KWCLPV | .NET Framework 4.7.2 | 49,466.96 μs | 981.576 μs | 2,462.582 μs | 49,396.60 μs | 43,798.700 μs | 54,853.00 μs |  327728 B |
| &#39;Calculate Delta - Remove lots of TagHelpers&#39; | Job-KWCLPV | .NET Framework 4.7.2 | 13,231.70 μs | 525.992 μs | 1,542.643 μs | 13,132.45 μs | 10,727.250 μs | 17,487.25 μs |  172080 B |
|     &#39;Calculate Delta - Mutate two TagHelpers&#39; | Job-KWCLPV | .NET Framework 4.7.2 | 25,743.97 μs | 617.644 μs | 1,762.174 μs | 25,712.95 μs | 21,305.300 μs | 29,829.20 μs |  311296 B |
|                 &#39;Calculate Delta - No change&#39; | Job-KWCLPV | .NET Framework 4.7.2 | 25,913.27 μs | 695.007 μs | 2,016.342 μs | 25,621.10 μs | 21,351.200 μs | 30,405.70 μs |  311296 B |

### New (this PR)

|                                        Method |        Job |            Toolchain |        Mean |      Error |     StdDev |      Median |          Min |         Max | Allocated |
|---------------------------------------------- |----------- |--------------------- |------------:|-----------:|-----------:|------------:|-------------:|------------:|----------:|
|               &#39;Calculate Delta - New project&#39; | Job-FWKRFG |             .NET 7.0 |    21.89 μs |   1.383 μs |   3.901 μs |    20.45 μs |    17.000 μs |    34.30 μs |         - |
|            &#39;Calculate Delta - Remove project&#39; | Job-FWKRFG |             .NET 7.0 |    18.68 μs |   1.468 μs |   4.093 μs |    17.65 μs |    13.250 μs |    32.15 μs |         - |
|    &#39;Calculate Delta - Add lots of TagHelpers&#39; | Job-FWKRFG |             .NET 7.0 | 1,679.28 μs | 169.092 μs | 479.686 μs | 1,608.30 μs |   858.600 μs | 3,165.50 μs |  234640 B |
| &#39;Calculate Delta - Remove lots of TagHelpers&#39; | Job-FWKRFG |             .NET 7.0 | 1,189.67 μs | 164.061 μs | 478.573 μs | 1,093.80 μs |   573.450 μs | 2,482.45 μs |  115224 B |
|     &#39;Calculate Delta - Mutate two TagHelpers&#39; | Job-FWKRFG |             .NET 7.0 | 1,350.06 μs | 181.911 μs | 527.758 μs | 1,233.90 μs |   713.800 μs | 2,812.20 μs |  147200 B |
|                 &#39;Calculate Delta - No change&#39; | Job-FWKRFG |             .NET 7.0 |   928.12 μs |  91.638 μs | 258.466 μs |   796.35 μs |   653.450 μs | 1,726.75 μs |  147056 B |
|               &#39;Calculate Delta - New project&#39; | Job-HXOHRU | .NET Framework 4.7.2 |    14.69 μs |   1.048 μs |   2.887 μs |    13.90 μs |    10.750 μs |    24.35 μs |         - |
|            &#39;Calculate Delta - Remove project&#39; | Job-HXOHRU | .NET Framework 4.7.2 |    11.21 μs |   0.759 μs |   2.116 μs |    10.40 μs |     8.100 μs |    17.60 μs |         - |
|    &#39;Calculate Delta - Add lots of TagHelpers&#39; | Job-HXOHRU | .NET Framework 4.7.2 | 1,863.16 μs | 170.603 μs | 475.572 μs | 1,756.60 μs | 1,138.900 μs | 3,330.70 μs |  238008 B |
| &#39;Calculate Delta - Remove lots of TagHelpers&#39; | Job-HXOHRU | .NET Framework 4.7.2 | 1,120.60 μs | 105.669 μs | 304.879 μs | 1,041.80 μs |   678.450 μs | 2,019.85 μs |  117792 B |
|     &#39;Calculate Delta - Mutate two TagHelpers&#39; | Job-HXOHRU | .NET Framework 4.7.2 | 1,536.71 μs | 137.174 μs | 386.902 μs | 1,476.60 μs |   913.000 μs | 2,676.30 μs |  148736 B |
|                 &#39;Calculate Delta - No change&#39; | Job-HXOHRU | .NET Framework 4.7.2 | 1,553.60 μs | 155.296 μs | 440.548 μs | 1,408.80 μs | 1,068.700 μs | 2,935.90 μs |  148736 B |